Overhaul administration scripts and cluster metadata storage.
Summary:
This contains several significant changes, the most of which is the
change to include database configuration as part of the group
configuration, rather than a separate structure. (This will
substantially simplify the consumer implementation.) As such, all
database-oriented commands have been removed and are now part of the
group management commands. (This also forces all of the database
management code to be reentrant, which is probably a good practice
anyway.)

This also changes the storage codec to use protocol buffers, rather than
JSON. All configuration commands that take payloads now use the protocol
buffer text format, and the storage in ZooKeeper uses the binary
encoding.

This also removes the consumer temporarily.

Reviewers: jeff, tail

Reviewed By: tail

Differential Revision: http://phabricator.local.disqus.net/D15456