Add snapshot function.
Summary:
This will be the function used by the consumer to take a consistent
snapshot of capture groups that span multiple tables.

I wanted to make it use a prepared statement, but couldn't figure out how to
make it work with psycopg2, so putting that off for the short term.

Test Plan:
I added a query tool for my own testing, and ended up just leaving
it in because it seems like it might be useful in the future. To use it, create
a file at $USER_TABLE with this as the contents:

    name: "auth_user"
    primary_key: "id"
    columns: "date_joined"
    columns: "email"
    columns: "password"
    columns: "username"
    joins {
        table {
            name: "authx_userprofile"
            primary_key: "id"
            columns: "about"
            columns: "avatar_url"
            columns: "display_name"
            columns: "flags"
            columns: "location"
            columns: "url"
        }
        foreign_key: "user_id"
    }

... and set $DATABASE to the analytics database DSN, then run:

    python src/main/python/pgshovel/snapshot.py $DATABASE $USER_TABLE 3064025 18987941 75642599 16872861 0

You sholud get the snapshots for myself, Dan, Jeff and Sam's user accounts.

Reviewers: danf, tail, jeff

Reviewed By: jeff

Differential Revision: http://phabricator.local.disqus.net/D15706