Make complex snapshots use the same date formatting as PostgreSQL.
Summary:
This ensures that trigger data and snapshot data is equivalent.

I'd rather make this happen in the trigger itself (so that the
transaction record would include a `datetime` object, rather than a
string representation of the date) but that would included pushing type
information into the trigger specification -- which is probably a good
idea, but not something I feel is necessarily good idea to tackle right
now (also see the similar rationale in {D16701}.)

Test Plan: Added unit test.

Reviewers: danf, tail, jeff

Reviewed By: jeff

Differential Revision: http://phabricator.local.disqus.net/D16874