Add vendor system, managed running of services
Summary:
Adds a `vendor` folder, with a `Makefile` to `make` vendored software.  At present, there are `make` targets for:

* `postgres`
* `zookeeper`
* `kafka`

Once vendored, one can use any of the "fixtures" in the `vendor/__init__.py` module to actually run those vendored pieces of software.  Each fixture has at a bare minimum, a `start()` and `stop()` method to start and stop the service.  The `postgres` fixture also ha a `createdb()` method to create a database.  One can use the `get_open_port()` helper function to find an open port to use, but I didn't want to complicate the fixtures with port finding code since I felt that was more a concern of the user of the fixture, not the fixture itself.

This code is basically a cleanup of [[http://phabricator.local.disqus.net/diffusion/EVENTS/browse/master/recorder/tests/fixtures.py | the same code in rEVENTS]], which itself was a borrowing of [[https://github.com/mumrah/kafka-python/blob/master/test/fixtures.py | code from `kafka-python`]].

The basic intention is that this vendor module is more-or-less tangential to the application code, and we could package it up separately as its own thing.  The actual fixture code is still kinda janky and feels not abstract enough, but I felt this was a good starting point.

Test Plan:
I ran this in a script:

```
vendor_root = os.path.abspath(os.path.dirname(__file__))

postgres = LocalPostgres(
    root=os.path.join(vendor_root, 'postgres', 'build'),
)

zookeeper = LocalZookeeper(os.path.join(vendor_root, 'zookeeper'))

with zookeeper:
    kafka = LocalKafka(
        root=os.path.join(vendor_root, 'kafka'),
        broker_id=12345,
        zk_host=zookeeper.host,
        zk_port=zookeeper.port
    )
    with kafka:
        with postgres:
            print 'We have reached the code'

print 'And we are done!'
```

And it worked.

Reviewers: ted

Reviewed By: ted

Differential Revision: http://phabricator.local.disqus.net/D17542